Address GCC integer truncation warning

+++ include/abaclade/enum.hxx
@@ -119,7 +119,7 @@ This design is loosely based on <http://www.python.org/dev/peps/pep-0435/>.
 #define _ABC_ENUM_MEMBER_ARRAY_ITEM(name) \
             { \
                ABC_SL(ABC_CPP_TOSTRING(name)), \
-               ABC_COUNTOF(ABC_CPP_TOSTRING(name)) - 1 /*NUL*/, \
+               static_cast<unsigned short>(ABC_COUNTOF(ABC_CPP_TOSTRING(name)) - 1 /*NUL*/), \
                name \
             },
